Updated photobooth code to email photo
Photobooth Processing sketch now runs once.  When called, it turns on
the camera, waits 5 seconds (delay for the camera to get initialized,
focused), and then takes a picture.  It then sends the picture as an
attachment to an email address.  The compiled version of the Processing
sketch is just a batch file that runs the code.  Had to edit the batch
file as the path to java.exe is not in the path variable.